+++ src/Toml/Edsl.hs
@@ -9,20 +9,28 @@ exampleToml = mkToml $ do
     \"key2\" =: Bool True
     table \"tableName\" $
         \"tableKey\" =: Array [\"Oh\", \"Hi\", \"Mark\"]
+    tableArray \"arrayName\" $
+        \"elem1\" =: \"yes\" :|
+        [ table \"elem2\" $ \"deep\" =: Integer 7
+        , empty
+        ]
 @
 
 -}
 
 module Toml.Edsl
        ( mkToml
+       , empty
        , (=:)
        , table
+       , tableArray
        ) where
 
-import Control.Monad.State (State, execState, modify)
+import Control.Monad.State (State, execState, modify, put)
+import Data.List.NonEmpty (NonEmpty)
 
 import Toml.PrefixTree (Key)
-import Toml.Type (TOML (..), Value, insertKeyVal, insertTable)
+import Toml.Type (TOML (..), Value, insertKeyVal, insertTable, insertTableArrays)
 
 
 type TDSL = State TOML ()
@@ -31,6 +39,10 @@ type TDSL = State TOML ()
 mkToml :: TDSL -> TOML
 mkToml env = execState env mempty
 
+-- | Creates an empty 'TDSL'.
+empty :: TDSL
+empty = put mempty
+
 -- | Adds key-value pair to the 'TDSL'.
 (=:) :: Key -> Value a -> TDSL
 (=:) k v = modify $ insertKeyVal k v
@@ -38,3 +50,7 @@ mkToml env = execState env mempty
 -- | Adds table to the 'TDSL'.
 table :: Key -> TDSL -> TDSL
 table k = modify . insertTable k . mkToml
+
+-- | Adds array of tables to the 'TDSL'.
+tableArray :: Key -> NonEmpty TDSL -> TDSL
+tableArray k = modify . insertTableArrays k . fmap mkToml